+++ Charcutarie.Api/Charcutarie.Application/Contracts/IOrderItemApp.cs
@@ -7,12 +7,12 @@ namespace Charcutarie.Application.Contracts
 {
     public interface IOrderItemApp
     {
-        Task<IEnumerable<OrderItem>> GetAll(int orderNumber, int corpClientId);
+        Task<IEnumerable<OrderItem>> GetAll(long orderNumber, int corpClientId);
         Task Remove(long orderId, long orderItemId, int corpClientId);
         Task Update(UpdateOrderItem model, int corpClientId);
         Task<long> AddOrderItem(NewOrderItem model, int corpClientId);
-        Task<int> GetLastItemNumber(int orderNumber, int corpClientId);
-        Task UpdateAllOrderItemStatus(int orderNumber, OrderItemStatusEnum status, int corpClientId);
+        Task<int> GetLastItemNumber(long orderNumber, int corpClientId);
+        Task UpdateAllOrderItemStatus(long orderNumber, OrderItemStatusEnum status, int corpClientId);
         Task<OrderItem> Get(long orderItemId, int corpClientId);
     }
 }
