+++ Charcutarie.Api/Charcutarie.Application/Contracts/IPricingApp.cs
@@ -5,7 +5,7 @@ namespace Charcutarie.Application.Contracts
 {
     public interface IPricingApp
     {
-        double CalculatePricePerTotalWeight(PriceRequest model, MeasureUnitTypeEnum pType, MeasureUnitTypeEnum qType);
-        double CalculatePricePerUnit(PriceRequest model, MeasureUnitTypeEnum pType, MeasureUnitTypeEnum qType);
+        decimal CalculatePricePerTotalWeight(PriceRequest model, MeasureUnitTypeEnum pType, MeasureUnitTypeEnum qType);
+        decimal CalculatePricePerUnit(PriceRequest model, MeasureUnitTypeEnum pType, MeasureUnitTypeEnum qType);
     }
 }
