+++ Charcutarie.Api/Charcutarie.Models/Entities/DataSheet.cs
@@ -11,7 +11,7 @@ namespace Charcutarie.Models.Entities
         public string ProcedureDescription { get; set; }
         public List<DataSheetItem> DataSheetItems { get; set; }
         public Product Product { get; set; }
-        public double WeightVariationPercentage { get; set; }
+        public decimal WeightVariationPercentage { get; set; }
         public bool IncreaseWeight { get; set; }
         
     }
