+++ Charcutarie.Api/Charcutarie.Models/Entities/ProfitSummary.cs
@@ -7,7 +7,7 @@ namespace Charcutarie.Models.Entities
     public class ProfitSummary
     {
         public Guid RowId { get; set; }
-        public double TotalProfit { get; set; }
-        public double CurrentMonthProfit { get; set; }
+        public decimal TotalProfit { get; set; }
+        public decimal CurrentMonthProfit { get; set; }
     }
 }
