+++ Charcutarie.Api/Charcutarie.Models/Entities/SalesPerMonth.cs
@@ -7,7 +7,7 @@ namespace Charcutarie.Models.Entities
     public class SalesPerMonth
     {
         public DateTime Date { get; set; }
-        public double TotalSales { get; set; }
-        public double TotalProfit { get; set; }
+        public decimal TotalSales { get; set; }
+        public decimal TotalProfit { get; set; }
     }
 }
