+++ Charcutarie.Api/Charcutarie.Models/Entities/SalesProfit.cs
@@ -7,7 +7,7 @@ namespace Charcutarie.Models.Entities
     public class SalesSummary
     {
         public Guid RowId { get; set; }
-        public double TotalSales { get; set; }
-        public double CurrentMonthSales { get; set; }
+        public decimal TotalSales { get; set; }
+        public decimal CurrentMonthSales { get; set; }
     }
 }
