+++ Charcutarie.Api/Charcutarie.Models/ViewModels/NewDataSheetItem.cs
@@ -8,7 +8,7 @@ namespace Charcutarie.Models.ViewModels
     {
         public long ProductId { get; set; }
         public long RawMaterialId { get; set; }
-        public double Percentage { get; set; }
+        public decimal Percentage { get; set; }
         public string AdditionalInfo { get; set; }
         public bool IsBaseItem { get; set; }
     }
