+++ Charcutarie.Api/Charcutarie.Models/ViewModels/PendingPaymentsSummary.cs
@@ -7,7 +7,7 @@ namespace Charcutarie.Models.ViewModels
     public class PendingPaymentsSummary
     {
         public Guid RowId { get; set; }
-        public double TotalPendingPayments { get; set; }
-        public double FinishedOrdersPendingPayment { get; set; }
+        public decimal TotalPendingPayments { get; set; }
+        public decimal FinishedOrdersPendingPayment { get; set; }
     }
 }
