+++ Charcutarie.Api/Charcutarie.Models/ViewModels/ProfitSummary.cs
@@ -7,7 +7,7 @@ namespace Charcutarie.Models.ViewModels
     public class ProfitSummary
     {
         public Guid RowId { get; set; }
-        public double TotalProfit { get; set; }
-        public double CurrentMonthProfit { get; set; }
+        public decimal TotalProfit { get; set; }
+        public decimal CurrentMonthProfit { get; set; }
     }
 }
