+++ Charcutarie.Api/Charcutarie.Services/Contracts/IPricingService.cs
@@ -4,7 +4,7 @@ namespace Charcutarie.Services.Contracts
 {
     public interface IPricingService
     {
-        double CalculatePricePerTotalWeight(PriceRequest model);
-        double CalculatePricePerUnit(PriceRequest model);
+        decimal CalculatePricePerTotalWeight(PriceRequest model);
+        decimal CalculatePricePerUnit(PriceRequest model);
     }
 }
